Fix import changes for file_utils
